requirements.txt
selenium
pandas
webdriver-manager
requests
streamlit
python-dotenv
plotly
streamlit-option-menu
